عملت ازرار المعيد وعملت التعديل على الجدول من دبل كليك وفي شغلة الي هي عملت اني بقدر احذف بيانات من الجدول عن طريق زر الdelete في الكيبورد بس المشكلة بنحفظش بتهيألي هيك اريح من انو نعمل زر التحرير دبل كليك تعديل وdelete للحذف بتقدر تشوف قصة الحذف في جدول المساقات بس لاني مكملتهاش عشان المشكلة وبس

+++ src/main/java/com/studentsattendance/models/Student.java
@@ -14,10 +14,10 @@ public class Student implements Serializable {
     private String major;
     private boolean status;
 
-    private int phoneNumber;
+    private long phoneNumber;
 
 
-    public Student(String firstName, String lastName, String email, String major, int phoneNumber) {
+    public Student(String firstName, String lastName, String email, String major, long phoneNumber) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
@@ -28,7 +28,7 @@ public class Student implements Serializable {
     //----------------------------------------------------------------
 // Constructors
 // ----------------------------------------------------------------
-    public Student(String username, String password, String firstName, String lastName, String email, String major, int phoneNumber) {
+    public Student(String username, String password, String firstName, String lastName, String email, String major, long phoneNumber) {
         this.username = username;
         this.password = password;
         this.firstName = firstName;
@@ -103,11 +103,11 @@ public class Student implements Serializable {
         this.status = status;
     }
 
-    public int getPhoneNumber() {
+    public long getPhoneNumber() {
         return phoneNumber;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
